move cpp extraction rules behind the indexer language registry
S3-CPP (ARCH-012), the sibling of S3-KOTLIN. C++'s tree-sitter rule file,
compiled-rule cache, member-call callsite marker, and receiver-call resolution
engine move out of the indexer god file into
`codestory-indexer/src/languages/cpp.rs`, as one `LanguageExtraction` row in
`languages/mod.rs`. Every language-keyed dispatch that spelled "cpp" now reads
the registry first and keeps its residual `match` for the fourteen languages
that have not moved; the last sibling package deletes those arms.

The move is mechanical: the 19 function bodies in `languages/cpp.rs` are
byte-identical to the ones deleted from `lib.rs` — an empty diff over 450 lines
— apart from renaming the two entry points
(`collect_cpp_receiver_call_edges` -> `receiver_call_specs`, `cpp_member_call`
-> `member_call`). Every helper moved because C++ was its only caller; the
shared seams stayed: `infer_cpp_access_from_tree` (C shares it),
`collect_cpp_declaration_span_overrides`,
`cpp_unknown_capture_needs_terminal_identifier`,
`collect_cpp_template_type_argument_edges`, and `CppSemanticResolver`.

`cpp_language_config` stays in `lib.rs` because it serves the `.h` header
seam — `h` is routed to `c` by the public registry, so that path cannot go
through `get_language_for_ext` — but it now builds its config from the C++
registry row rather than from its own copy of the rule file.

Two roster values are deliberately not "tidied" to match C++ syntax, because
they are separate facts with separate owners:

* `route_comments_are_c_style: false`. The framework-route text scanner's
  roster never listed `cpp`. It is a claim about which languages carry
  scannable routes, not a syntax fact.
* `promotes_type_member_functions_to_methods: false`. The promotion roster held
  only `kotlin`, `swift`, and `dart`; C++ member functions project as FUNCTION.

Proof: `tests/language_extraction_snapshot.rs` gains one row plus two goldens
pinning the full rendered projection of both C++ fixtures — the fidelity_lab
file through the store pipeline (resolution included) and the tictactoe file
through raw `index_file`. Both goldens were captured before the production edit
and are byte-identical after (sha256 9570a363.. / 79a71fd3..). Mutation-tested:
flipping `promotes_type_member_functions_to_methods` or changing
`qualified_name_delimiter` from `::` to `.` fails these two snapshots and
nothing else in the crate — `fidelity_regression` and
`tictactoe_language_coverage` both stay green.

Declared behavior changes:
- None to any indexing, resolution, or rendering output. C++
  fidelity_regression and tictactoe_language_coverage projections are proven
  byte-identical before and after, and packet_search_eval is unmoved.
- `LanguageRuleset::Cpp`'s arm in `compiled_rules` is now unreachable and
  returns a typed error instead of compiled rules; the registry answers first.
  Fail-closed by construction, matching the `Kotlin` arm.
- `crate::CPP_MEMBER_CALLSITE_MARKER` becomes
  `crate::languages::cpp::MEMBER_CALLSITE_MARKER`; same value, same single
  consumer in `resolution`.
- `codestory_contracts::language_support::LANGUAGE_COMMENT_PROFILES` gains a
  `cpp` row and the CLI highlight roster drops its `cpp` arm. Same marker,
  same rendering; `comment_markers_stay_identical_after_the_registry_lookup`
  restates the pre-move roster and still passes unchanged.
- The CLI comment roster and the indexer route-comment roster still disagree
  (CLI has cpp/swift/tsx/jsx and not c; the route scanner has vue/astro and not
  cpp/swift/tsx). That divergence is preserved, not unified.
- No version surface touched; everything stays 0.16.3.

Closes #1678
