+++ crates/codestory-cli/src/output.rs
@@ -3942,7 +3942,7 @@ fn ansi_highlight_line(language: &str, line: &str) -> String {
         .or(match language {
             "bash" | "python" | "ruby" | "toml" | "yaml" => Some("#"),
             "rust" | "typescript" | "tsx" | "javascript" | "jsx" | "go" | "java" | "csharp"
-            | "cpp" | "dart" | "php" | "swift" => Some("//"),
+            | "dart" | "php" | "swift" => Some("//"),
             _ => None,
         });
     let Some(marker) = comment_marker else {
@@ -5639,6 +5639,14 @@ mod tests {
         );
         let kotlin = ansi_highlight_snippet("app/Main.kt", "val ok = true // comment");
         assert!(kotlin.contains("\x1b[90m// comment\x1b[0m"), "{kotlin:?}");
+
+        // Same for C++, whose roster row moved into the registry with S3-CPP.
+        assert_eq!(
+            codestory_contracts::language_support::line_comment_for_language("cpp"),
+            Some("//")
+        );
+        let cpp = ansi_highlight_snippet("src/main.cpp", "int ok = 1; // comment");
+        assert!(cpp.contains("\x1b[90m// comment\x1b[0m"), "{cpp:?}");
     }
 
     /// Component dialects resolve through the companion-extension registry and
