+++ crates/codestory-indexer/src/resolution/mod.rs
@@ -1321,7 +1321,7 @@ fn is_cpp_member_call_placeholder(edge_kind: EdgeKind, callsite_identity: Option
     callsite_identity.is_some_and(|identity| {
         identity
             .split('|')
-            .any(|part| part == crate::CPP_MEMBER_CALLSITE_MARKER)
+            .any(|part| part == crate::languages::cpp::MEMBER_CALLSITE_MARKER)
     })
 }
